Improve consistency when refers to symlinks
"symlink" should be a well understood term for users of this library.
Previous attemts to "English" the API is dubious at best.

However, due to the fact that these changes happens at arugment label
level, all of which came with default values, we can not properly
deprecate the old function names without introducing ambiguities at
user's callsite. So these will have to be breaking changes.